+++ src/app/[locale]/topics/[scope]/[topicId]/edit/EditTopicPage.tsx
@@ -373,7 +373,7 @@ export function EditTopicPage(props: TEditTopicPageProps) {
         content: t('Save'),
         variant: 'success',
         icon: Icons.Save,
-        visibleFor: 'sm',
+        visibleFor: 'xs',
         pending: isPending,
         disabled: !isSubmitEnabled,
         onClick: handleSubmit,