+++ src/features/ai-generations/constants/index.ts
@@ -1,4 +1,4 @@
 import { isDev } from '@/config';
 
-export const maxAnswersToGeneration = isDev ? 10 : 20;
-export const maxQuestionsToGeneration = isDev ? 10 : 20;
+export const maxAnswersToGeneration = isDev ? 10 : 10;
+export const maxQuestionsToGeneration = isDev ? 10 : 10;