+++ src/features/ai/types/TAIQueryOptions.ts
@@ -1,6 +1,10 @@
 import { TAiClientType } from '@/lib/ai/types/TAiClientType';
 
-export type TAIQuerDebugDataId = 'answers-query-data-01' | 'questions-query-data-01' | boolean;
+export type TAIQuerDebugDataId =
+  | 'answers-query-data-01'
+  | 'questions-query-data-01'
+  | 'questions-query-data-02'
+  | boolean;
 
 export interface TAIQueryOptions {
   clientType?: TAiClientType;