+++ src/features/bot/actions/sendLoggingMessage.ts
@@ -4,14 +4,16 @@ import { LOGGING_CHANNEL_ID } from '@/config/envServer';
 import { getErrorText } from '@/lib/helpers';
 import { getBot } from '@/features/bot/core/getBot';
 
+const limit = 4000; // Keep slightly below 4096 for safety
+
 export async function sendLoggingMessage(text: string) {
   try {
     const bot = getBot();
-    /* console.log('[sendLoggingMessage]', {
-     *   text,
-     * });
-     */
-    await bot.api.sendMessage(LOGGING_CHANNEL_ID, text);
+    // Send large messages splitted
+    for (let i = 0; i < text.length; i += limit) {
+      const part = text.substring(i, i + limit);
+      await bot.api.sendMessage(LOGGING_CHANNEL_ID, part);
+    }
   } catch (error) {
     const errMsg = getErrorText(error);
     // eslint-disable-next-line no-console