+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 <!--
  @since 2025.07.05
- @changed 2026.01.31, 22:30
+ @changed 2026.02.01, 22:37
 -->
 
 # CHANGELOG
+
+## Fixed minor issues - 2026.02.01
+
+- Fixed an error with too strict expectations for generated answers (categoryIds and isCorrect flags became optional).
+- Added a function to send logs to the telegram bot (`logData`).
+- Added demo data with problematic generation results (`questions-query-data-02`).
+- Updated questions and answers count to generate constants.
+- Added an alternative way to start nextjs dev server.
+- Added splitting large telegram messages.
 
 ## [Issue #68: Updated logo, translations, pricing calculations, and routing](https://github.com/lilliputten/mindstack/issues/68) - 2026.01.31